Moved logic from page.svelte to component and moved gpaphviz related logic to its own file to avoid code duplication.
